move UTs to separate files, decide on membrane from environment, add reporter
