+++ .env
@@ -1,0 +1,1 @@
+JEST_MEMBRANE=6 7 8
