+++ .gitignore
@@ -1,2 +1,4 @@
 stash
 node_modules
+jest_html_reporters
+jest_html_reporters.html
