+++ jest.config.js
@@ -1,5 +1,8 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
+  preset: "ts-jest",
+  testEnvironment: "node",
+  setupFilesAfterEnv: ["./jest.setup.ts"],
+  reporters: ["default", "jest-html-reporters"],
+  verbose: true,
 };
