+++ package.json
@@ -9,9 +9,12 @@
     "test": "jest"
   },
   "devDependencies": {
+    "@types/dotenv": "^8.2.0",
     "@types/jest": "^29.5.11",
+    "dotenv": "^16.3.1",
     "http-server": "^14.1.1",
     "jest": "^29.7.0",
+    "jest-html-reporters": "^3.1.7",
     "ts-jest": "^29.1.1",
     "typescript": "^5.3.3"
   }
